Refactor tabs into separate modules and fix inspector
