+++ src/gui/__init__.py
@@ -1,2 +1,5 @@
 """GUI components for the Automation Recorder application."""
 
+from .automation_recorder import AutomationRecorderApp
+
+__all__ = ["AutomationRecorderApp"]
